Writing directory to file system is the last operation.
Within the NewDir method of cloudstore.IO, writing the directory needs to be the last operation. If any query method fails the database is rolled back. The directory is only removed from file system if commiting the operation fails, which is marked by ErrCommitTx. If any queries were to fail after writing the directory, it would not return an ErrCommitTx and there would be a rogue directory.