Created POST /api/upload handler.
Functionality not yet implemented.

UploadPath handler gets the path using the "path" URL query parameter, and will use that to save the files. The path will be the user-facing directory path, not the UUID paths that are stored on the server. Each directory in the path should reference a directory name in the database.